Reorganização dos Packages da aplicação
