+++ src/main/java/com/br/zup/vicente_imoveis/Cliente/ClienteRepository.java
@@ -1,7 +1,9 @@
-package com.br.zup.vicente_imoveis.cliente;
+package com.br.zup.vicente_imoveis.Cliente;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClienteRepository extends CrudRepository<Cliente, String> {
+import java.util.List;
 
+public interface ClienteRepository extends CrudRepository<Cliente, String> {
+    List<Cliente> findAllByCpf (String cpf);
 }
