+++ src/main/java/com/br/zup/vicente_imoveis/Contrato/Dtos/ContratoSaidaDTO.java
@@ -1,8 +1,8 @@
-package com.br.zup.vicente_imoveis.Contrato.dto;
+package com.br.zup.vicente_imoveis.Contrato.Dtos;
 
-import com.br.zup.vicente_imoveis.Contrato.enun.StatusDoContrato;
+import com.br.zup.vicente_imoveis.Contrato.Enums.StatusDoContrato;
 import com.br.zup.vicente_imoveis.Imovel.Imovel;
-import com.br.zup.vicente_imoveis.cliente.Cliente;
+import com.br.zup.vicente_imoveis.Cliente.Cliente;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
