+++ src/main/java/com/br/zup/vicente_imoveis/Contrato/Enums/StatusDoContrato.java
@@ -1,4 +1,4 @@
-package com.br.zup.vicente_imoveis.Contrato.enun;
+package com.br.zup.vicente_imoveis.Contrato.Enums;
 
 public enum StatusDoContrato {
     ATIVO,
