+++ src/main/java/com/br/zup/vicente_imoveis/Imovel/Dtos/ImovelAtualizarDTO.java
@@ -1,4 +1,4 @@
-package com.br.zup.vicente_imoveis.Imovel;
+package com.br.zup.vicente_imoveis.Imovel.Dtos;
 
 import com.br.zup.vicente_imoveis.Imovel.Enums.TipoDeContrato;
 import lombok.Getter;
